+++ src/main/kotlin/icu/windea/pls/core/psi/ParadoxPsiManager.kt
@@ -8,8 +8,6 @@ import icu.windea.pls.localisation.psi.*
 object ParadoxPsiManager {
     /**
      * 判断当前位置应当是一个[ParadoxLocalisationLocale]，还是一个[ParadoxLocalisationPropertyKey]。
-     *
-     * 优化代码提示时会用到此方法。
      */
     fun isLocalisationLocaleLike(element: PsiElement): Boolean {
         val elementType = element.elementType