+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/completion/ParadoxLocalisationPropertyReferenceCompletionProvider.kt
@@ -40,7 +40,7 @@ class ParadoxLocalisationPropertyReferenceCompletionProvider : CompletionProvide
             .contextSensitive()
             .preferLocale(ParadoxLocaleHandler.getPreferredLocale())
             //.distinctByName() //这里selector不需要指定去重
-        val processor: (ParadoxLocalisationProperty) -> Boolean = {
+        val processor = LimitedCompletionProcessor<ParadoxLocalisationProperty> {
             ProgressManager.checkCanceled()
             val name = it.name
             val icon = it.icon